Make executeWorkbench work after VSCode restarts
`vscode.openFolder` can be used to restart the VSCode under different
workspace root. In that case the current socket is closed, and then the new
connection appears after the VSCode is up again.

This PR clears pending message callbacks when the connection is closed, and also
re-assigns the `_promisedSocket` to a new socket when new connection appears.